requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
xgboost>=1.6
mlflow>=2.3
google-cloud-storage>=2.10
google-cloud-aiplatform>=1.38
kfp>=2.4.0
pytest>=7.0
db-dtypes>=1.1.1
